--- a/Plugins/Replicas/Source/Replicas/Private/Replicas.cpp
+++ b/Plugins/Replicas/Source/Replicas/Private/Replicas.cpp
@@ -1,7 +1,6 @@
 // Copyright 1998-2018 Epic Games, Inc. All Rights Reserved.
 
 #include "Replicas.h"
-#include "ReplicasPrivatePCH.h"
 
 #define LOCTEXT_NAMESPACE "FReplicasModule"