--- a/Plugins/Replicas/Source/Replicas/Public/Replicas.h
+++ b/Plugins/Replicas/Source/Replicas/Public/Replicas.h
@@ -2,7 +2,6 @@
 
 #pragma once
 
-//#include "CoreMinimal.h"
 #include "ModuleManager.h"
 
 class FReplicasModule : public IModuleInterface