--- a/Source/RN4UE4/Tests/BoundaryBox.cpp
+++ b/Source/RN4UE4/Tests/BoundaryBox.cpp
@@ -1,7 +1,6 @@
 // Fill out your copyright notice in the Description page of Project Settings.
 
 #include "BoundaryBox.h"
-#include "RakNetRP.h"
 
 ABoundaryBox::ABoundaryBox()
 {