--- a/Source/RN4UE4/Tests/SpawnVolume.cpp
+++ b/Source/RN4UE4/Tests/SpawnVolume.cpp
@@ -1,9 +1,6 @@
 // Fill out your copyright notice in the Description page of Project Settings.
 
 #include "SpawnVolume.h"
-#include "Rand.h"
-#include "EngineUtils.h"
-#include "../RakNetRP.h"
 #include "RN4UE4GameInstance.h"
 
 ASpawnVolume::ASpawnVolume()