--- a/Source/RN4UE4/Tests/SpawnVolume.h
+++ b/Source/RN4UE4/Tests/SpawnVolume.h
@@ -5,6 +5,7 @@
 #include "CoreMinimal.h"
 #include "Components/BoxComponent.h"
 #include "GameFramework/Actor.h"
+#include "Math/RandomStream.h"
 #include "RakNetRP.h"
 #include "SpawnVolume.generated.h"