More build time optimisations.

Before:
```
1>Deploying RN4UE4Editor Win64 Development...
1>Total build time: 22.28 seconds (Local executor: 11.39 seconds)
========== Rebuild All: 1 succeeded, 0 failed, 0 skipped ==========
```

After:
```
1>Deploying RN4UE4Editor Win64 Development...
1>Total build time: 16.87 seconds (Local executor: 5.95 seconds)
========== Rebuild All: 1 succeeded, 0 failed, 0 skipped ==========
```